ace2: New character device driver for SIFT-PACMAN
This provides /dev/ace2-data and /dev/ace2-capability devices which
emulate the same interface as the /proc/ace2-data and
/proc/ace2-capability devices from SIFT-PACMAN/linux.

The file offset for both devices is a kernel address similar to
/dev/kmem and access is validated the same way using VIRT_IN_DMAP()
and kernacc().  /dev/ace2-data permits reading/writing scalar data.
/dev/ace2-capability treats the file offset as a starting virtual
address and returns the address of consecutive pointers.  Writes to
/dev/ace2-capability write the supplied addresses to existing
capabilities using cheri_setaddress preserving existing capability
metadata (and potentially clearing the tag if the new address is
unrepresentable).